+++ src/SymMFEA/evolution/ranker.py
@@ -3,7 +3,7 @@ import numba as nb
 from ..utils.timer import *
 from pygmo import sort_population_mo as mo_sort
 import os
-@nb.njit(nb.int64[:](nb.float64[:]), cache= os.environ.get('DISABLE_NUMBA_CACHE') is None)
+@nb.njit(nb.int64[:](nb.float32[:]), cache= os.environ.get('DISABLE_NUMBA_CACHE') is None)
 def sort_scalar_fitness(ls_fcost):
     return np.argsort(-ls_fcost)
